Refactor: Split SyncCard.tsx into smaller components
Split SyncCard.tsx into smaller, more manageable components to improve code organization and readability.
