+++ src/components/admin/sync-card/index.ts
@@ -1,0 +1,3 @@
+
+export * from "./SyncCard";
+export * from "./types";
